Genrative AI langchain/requirements.txt
ipykernel
langchain
python-dotenv
langchain-google-genai
langchain_google_genai
beautifulsoup4
langchain_community
faiss-cpu
streamlit